refactor: reroute components
