+++ src/components/card/UserInterviewSessionCard.tsx
@@ -24,7 +24,11 @@ export const UserInterviewSessionCard: React.FC<
       <DropdownMenuItem disabled={!onEdit} onClick={onEdit}>
         Edit Session
       </DropdownMenuItem>
-      <DropdownMenuItem disabled={!onDelete} onClick={onDelete}>
+      <DropdownMenuItem
+        variant="destructive"
+        disabled={!onDelete}
+        onClick={onDelete}
+      >
         Delete Session
       </DropdownMenuItem>
     </DropdownMenuContent>
